adding fragment currency data and back bug fixed

+++ wallet/app/src/main/java/com/github/derleymad/lizwallet/model/ListOfCurrencies.kt
@@ -9,7 +9,7 @@ data class ListOfCurrencies(
     val image: String,
     val current_price: Double,
     val market_cap: Double,
-    val market_cap_rank: Double,
+    val market_cap_rank: Int,
     val fully_diluted_valuation: Double,
     val total_volume: Double,
     val high_24h: Double,
